+++ lib/runners/harness.rb
@@ -113,7 +113,6 @@ module Runners
     end
 
     def exclude_special_dirs
-      # @type var saved: Hash<Pathname, Pathname>
       saved = [".git"].filter_map do |dir|
         src = working_dir / dir
         if src.directory?