+++ lib/runners/testing/smoke.rb
@@ -298,6 +298,7 @@ module Runners
             type: type,
             guid: guid,
             timestamp: timestamp,
+            issue_count: issues == :_ ? :_ : issues&.size,
             issues: issues,
             message: message,
             analyzer: analyzer,